requirements.txt
attrs>=17.4.0
aws-cdk-lib==2.59.0
constructs>=10.0.0
cattrs>=1.8
certifi==2020.6.20
chardet==3.0.4
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.0.0
iniconfig==1.0.1
jsii>=1.72.0
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
publication==0.0.3
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.2
urllib3==1.25.10
zipp==3.1.0
aws_lambda_powertools
types-boto3
types-setuptools
